+++ feed_renderer.py
@@ -5,36 +5,45 @@
 
 from __future__ import annotations
 
-from typing import Iterable, Dict, Any
+from typing import Iterable, Dict, Any, List
+from dataclasses import dataclass
+import random
 
 import streamlit as st
 
 from streamlit_helpers import render_post_card
 
-# Demo posts used when none are provided
-DEMO_POSTS: list[dict[str, Any]] = [
-    {
-        "image": "https://placekitten.com/400/300",
-        "text": "Look at this cute kitten!",
-        "likes": 5,
-    },
-    {
-        "image": "https://placekitten.com/500/300",
-        "text": "Another adorable cat appears.",
-        "likes": 3,
-    },
-    {
-        "image": "https://placekitten.com/450/300",
-        "text": "Cats everywhere!",
-        "likes": 8,
-    },
-]
+
+@dataclass
+class Post:
+    """Simple post structure for demo feeds."""
+
+    username: str
+    image: str
+    caption: str
+    reactions: Dict[str, int]
+
+
+def generate_demo_posts() -> List[Dict[str, Any]]:
+    """Return a few placeholder posts with random reactions."""
+    users = ["Alice", "Bob", "Carol"]
+    demo: List[Dict[str, Any]] = []
+    for idx, user in enumerate(users, start=1):
+        demo.append(
+            {
+                "username": user,
+                "image": f"https://placehold.co/600x400?text=Post+{idx}",
+                "text": f"Demo caption {idx}",
+                "likes": random.randint(1, 9),
+            }
+        )
+    return demo
 
 
 def render_feed(posts: Iterable[Dict[str, Any]] | None = None) -> None:
     """Render each post using :func:`render_post_card`."""
     if posts is None:
-        posts = DEMO_POSTS
+        posts = generate_demo_posts()
 
     posts = list(posts)
     if not posts:
@@ -43,3 +52,6 @@ def render_feed(posts: Iterable[Dict[str, Any]] | None = None) -> None:
 
     for post in posts:
         render_post_card(post)
+
+__all__ = ["render_feed", "generate_demo_posts", "Post"]
+
